rename has3DTilesExtension -> hasExtension
