+++ Source/Scene/ImplicitTileset.js
@@ -3,7 +3,7 @@ import clone from "../Core/clone.js";
 import defined from "../Core/defined.js";
 import Resource from "../Core/Resource.js";
 import RuntimeError from "../Core/RuntimeError.js";
-import has3DTilesExtension from "./has3DTilesExtension.js";
+import hasExtension from "./hasExtension.js";
 import ImplicitSubdivisionScheme from "./ImplicitSubdivisionScheme.js";
 
 /**
@@ -67,7 +67,7 @@ export default function ImplicitTileset(
   if (
     !defined(tileJson.boundingVolume.box) &&
     !defined(tileJson.boundingVolume.region) &&
-    !has3DTilesExtension(tileJson.boundingVolume, "3DTILES_bounding_volume_S2")
+    !hasExtension(tileJson.boundingVolume, "3DTILES_bounding_volume_S2")
   ) {
     throw new RuntimeError(
       "Only box, region and 3DTILES_bounding_volume_S2 are supported for implicit tiling"
@@ -219,7 +219,7 @@ export default function ImplicitTileset(
  * @private
  */
 function gatherContentHeaders(tileJson) {
-  if (has3DTilesExtension(tileJson, "3DTILES_multiple_contents")) {
+  if (hasExtension(tileJson, "3DTILES_multiple_contents")) {
     return tileJson.extensions["3DTILES_multiple_contents"].content;
   }
